create InvalidSessionHandler component
