+++ src/app/[locale]/clientLayout.tsx
@@ -9,17 +9,22 @@ import { Internationalization } from 'lib/i18n/Internationalization';
 import { ReactNode } from 'react';
 import { SessionProvider } from 'next-auth/react';
 import { EnvProvider } from 'app/EnvProvider';
+import { InvalidSessionHandler } from 'components/authentication/InvalidSessionHandler';
 
 export type ClientLayoutProps = {
     children: ReactNode;
 };
 
-const ONE_HOUR = 1000 * 60 * 60;
+// When setting up Mnestix in a kiosk mode it will run long without interaction
+// for this case we refetch the session every 5 minutes.
+// This does not refresh the session but will trigger a logout if the session is invalid.
+const FIVE_MIN = 5 * 60;
 
 export const ClientLayout = ({ children }: Readonly<ClientLayoutProps>) => {
     return (
         <EnvProvider>
-            <SessionProvider refetchInterval={ONE_HOUR} refetchOnWindowFocus={true}>
+            <SessionProvider refetchInterval={FIVE_MIN} refetchOnWindowFocus={true}>
+                <InvalidSessionHandler />
                 <Internationalization>
                     <CustomThemeProvider>
                         <CurrentAasContextProvider>
