+++ src/layout/LayoutRoot.tsx
@@ -1,13 +1,9 @@
 'use client';
-import { useTranslations } from 'use-intl';
 import { Box, styled } from '@mui/material';
 import { NotificationOutlet } from 'components/basics/NotificationOutlet';
 import { Header } from './Header';
 import { Footer } from './Footer';
-import { ReactNode, useEffect, useState } from 'react';
-import { Session } from 'next-auth';
-import { useNotificationSpawner } from 'lib/hooks/UseNotificationSpawner';
-import { useSession } from 'next-auth/react';
+import { ReactNode } from 'react';
 
 const StyledBox = styled(Box)(() => ({
     // toolbar min-height is 56px
@@ -22,27 +18,10 @@ type Props = {
     children: ReactNode;
 };
 
+/**
+ * Client side root render layout.
+ */
 export function LayoutRoot({ children }: Props) {
-    const [prefSession, setPrefSession] = useState<Session | null>(null);
-    const { data: session } = useSession();
-    const t = useTranslations('validation.authentication');
-
-    const notificationSpawner = useNotificationSpawner();
-
-    useEffect(() => {
-        if (prefSession && !session) {
-            showSessionExpired();
-        }
-        setPrefSession(session);
-    }, [session]);
-
-    function showSessionExpired() {
-        notificationSpawner.spawn({
-            message: t('sessionExpired'),
-            severity: 'warning',
-        });
-    }
-
     return (
         <Box display="flex" height="100%" flexDirection="column">
             <Box display="flex" flex={1} flexDirection="column">
